+++ term2/task3/salesman/check.cpp
@@ -3,14 +3,14 @@
 //
 #include "graph.h"
 
-int check(ListGraph graph) {
+float check(ListGraph graph) {
     vector<int> cities(graph.verticesCount);
     for (int i = 0; i < cities.size(); ++i) {
         cities[i] = i;
     }
-    int minPath = numeric_limits<int>::max();
+    float minPath = numeric_limits<int>::max();
     do {
-        int currentPath = 0;
+        float currentPath = 0;
         for (int i = 1; i < cities.size(); ++i) {
             currentPath += graph.weightMatrix[cities[i]][cities[i-1]];
         }
