+++ term2/task3/salesman/graph.cpp
@@ -5,7 +5,7 @@
 #include "graph.h"
 
 ListGraph::ListGraph(int vertexCount) : verticesCount(vertexCount),
-                                        weightMatrix(vertexCount, vector<int>(vertexCount, std::numeric_limits<int>::max())),
+                                        weightMatrix(vertexCount, vector<float>(vertexCount, std::numeric_limits<int>::max())),
                                         adjacencyList(vertexCount){
     for (int i = 0; i < weightMatrix.size(); ++i) {
         weightMatrix[i][i] = 0;
@@ -15,7 +15,7 @@ ListGraph::ListGraph(int vertexCount) : verticesCount(vertexCount),
 //ListGraph::ListGraph(const ListGraph& graph) : verticesCount(graph.verticesCount), edgesList(graph.edgesList) {}
 
 
-void ListGraph::addEdge(int from, int to, long weight) {
+void ListGraph::addEdge(int from, int to, float weight) {
     Edge edge(from, to, weight);
     adjacencyList[from].push_back(edge);
     Edge backEdge(to, from, weight);
